+++ src/log.c
@@ -287,7 +287,7 @@ Log_Print (EtLogAreaKind error_type, const gchar * const format, ...)
     va_list args;
     gchar *string;
     gchar *time;
-    gint n_items;
+    guint n_items;
     GtkTreeIter iter;
     static gboolean first_time = TRUE;
     static gchar *file_path = NULL;
@@ -308,8 +308,8 @@ Log_Print (EtLogAreaKind error_type, const gchar * const format, ...)
     time = Log_Format_Date ();
 
     /* Remove lines that exceed the limit. */
-    n_items = gtk_tree_model_iter_n_children (GTK_TREE_MODEL (priv->log_model),
-                                              NULL);
+    n_items = (guint)gtk_tree_model_iter_n_children (GTK_TREE_MODEL (priv->log_model),
+                                                     NULL);
 
     if (n_items > g_settings_get_uint (MainSettings, "log-lines") - 1
         && gtk_tree_model_get_iter_first (GTK_TREE_MODEL (priv->log_model),
